Move the inheriting IntersectionPredicate classes to a new header.
