+++ source/interaction/interaction_base.cc
@@ -1,6 +1,7 @@
 #include <fiddle/base/samrai_utilities.h>
 
 #include <fiddle/grid/box_utilities.h>
+#include <fiddle/grid/intersection_predicate_lib.h>
 
 #include <fiddle/interaction/interaction_base.h>
