+++ tests/grid/patch_map_02.cc
@@ -1,4 +1,5 @@
 #include <fiddle/grid/box_utilities.h>
+#include <fiddle/grid/intersection_predicate_lib.h>
 #include <fiddle/grid/overlap_tria.h>
 #include <fiddle/grid/patch_map.h>
